requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Deprecated==1.2.13
Django==3.2.6
django-appconf==1.0.5
django-bootstrap-datepicker-plus==3.0.5
django-cleanup==5.2.0
django-debug-toolbar==3.2.4
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-pwa==1.0.10
django-redis==5.2.0
django-select2==7.10.0
django-webpush==0.3.4
django-widget-tweaks==1.4.8
http-ece==1.1.0
idna==3.3
jsonfield==3.1.0
packaging==21.3
py-vapid==1.8.2
pycparser==2.21
pymemcache==3.5.1
pyparsing==3.0.7
pytz==2021.1
pywebpush==1.9.4
redis==4.2.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
swapper==1.3.0
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
